requirements/base.txt
aioredis==1.3.1
alabaster==0.7.12
appdirs==1.4.3
appnope==0.1.0
argon2-cffi==19.2.0
asgiref==3.2.7
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.4.1
Automat==20.2.0
Babel==2.8.0
backcall==0.1.0
beautifulsoup4==4.9.0
black==19.10b0
bleach==3.1.4
certifi==2019.11.28
cffi==1.14.0
cfgv==3.1.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
click==7.1.1
click-plugins==1.1.1
cligj==0.5.0
constantly==15.1.0
coverage==5.0.4
cryptography==2.8
daphne==2.4.1
dash==1.9.1
dash-core-components==1.8.1
dash-html-components==1.0.2
dash-renderer==1.2.4
dash-table==4.6.1
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
Django==3.0.4
django-allauth==0.41.0
django-appconf==1.0.4
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==1.1.1
django-cors-headers==3.2.1
django-coverage-plugin==1.8.0
django-crispy-forms==1.9.0
django-debug-toolbar==2.2
django-environ==0.4.5
django-extensions==2.2.9
django-extra-views==0.13.0
django-filter==2.2.0
django-geojson==3.0.0
django-ipware==2.1.0
django-leaflet==0.26.0
django-model-utils==4.0.0
django-phonenumber-field==4.0.0
django-plotly-dash==1.1.5
django-redis==4.11.0
django-reverse-admin==2.8.3
django-select2==7.2.3
django-sslserver==0.22
django-stubs==1.5.0
django-tables2==2.3.1
djangorestframework==3.11.0
djangorestframework-gis==0.15
docutils==0.16
dpd-components==0.1.0
entrypoints==0.3
factory-boy==2.12.0
Faker==4.0.2
filelock==3.0.12
Fiona==1.8.13.post1
flake8==3.7.9
flake8-isort==2.9.1
Flask==1.1.1
Flask-Compress==1.4.0
future==0.18.2
geopandas==0.7.0
gunicorn==20.0.4
hiredis==1.0.1
hyperlink==19.0.0
identify==1.4.13
idna==2.9
imagesize==1.2.0
importlib-metadata==1.6.0
importlib-resources==1.4.0
incremental==17.5.0
ipdb==0.13.2
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
jsonfield==3.1.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.2
jupyter-console==6.1.0
jupyter-core==4.6.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.2.0
msgpack==0.6.2
munch==2.5.0
mypy==0.770
mypy-extensions==0.4.3
nbconvert==5.6.1
nbformat==5.0.5
nodeenv==1.3.5
notebook==6.0.3
numpy==1.18.2
oauthlib==3.1.0
packaging==20.3
pandas==1.0.3
pandocfilters==1.4.2
parso==0.6.2
pathspec==0.7.0
pexpect==4.8.0
phonenumberslite==8.12.2
pickleshare==0.7.5
Pillow==7.0.0
plotly==4.5.4
pluggy==0.13.1
pre-commit==2.2.0
prometheus-client==0.7.1
prompt-toolkit==3.0.5
psycopg2-binary==2.8.4
ptyprocess==0.6.0
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.20
pyflakes==2.1.1
Pygments==2.6.1
PyHamcrest==2.0.2
pylint==2.4.4
pylint-django==2.0.14
pylint-plugin-utils==0.6
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyproj==2.6.0
pyrsistent==0.16.0
pytest==5.3.5
pytest-django==3.9.0
pytest-sugar==0.9.2
python-dateutil==2.8.1
python-slugify==4.0.0
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.3.1
pyzmq==19.0.0
qtconsole==4.7.2
QtPy==1.9.0
redis==3.4.1
regex==2020.2.20
requests==2.23.0
requests-oauthlib==1.3.0
retrying==1.3.3
Send2Trash==1.5.0
service-identity==18.1.0
Shapely==1.7.0
six==1.14.0
snowballstemmer==2.0.0
soupsieve==2.0
Sphinx==2.4.4
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.1
termcolor==1.1.0
terminado==0.8.3
testfixtures==6.14.0
testpath==0.4.4
text-unidecode==1.3
toml==0.10.0
tornado==6.0.4
traitlets==4.3.3
Twisted==20.3.0
txaio==20.4.1
typed-ast==1.4.1
typing-extensions==3.7.4.1
urllib3==1.25.8
virtualenv==20.0.15
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.0
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==3.1.0
zope.interface==5.0.2
